fix class
