+++ program/settings.json
@@ -1,4 +1,4 @@
-{
+﻿{
     "FilePath": {
         "StudentsListPath": "../生徒名簿.xlsx"
     },
@@ -15,11 +15,11 @@
     "Year": {
         "小2": "01",
         "小3": "07",
-        "小4": "34",
-        "小5": "22",
-        "小6": "20",
-        "中1": "21",
-        "中2": "25",
+        "小4": "20",
+        "小5": "21",
+        "小6": "22",
+        "中1": "25",
+        "中2": "34",
         "中3": "13"
     },
     "APP": {
